+++ payjoin/src/receive/error.rs
